requirements.txt
absl-py==0.3.0
astor==0.7.1
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
gast==0.2.0
graphviz==0.8.4
grpcio==1.14.1
idna==2.6
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.3
more-itertools==4.3.0
mxboard==0.1.0
mxnet-cu80==1.2.1
numpy==1.14.5
pathlib2==2.3.2
Pillow==5.2.0
pkg-resources==0.0.0
pluggy==0.7.1
protobuf==3.6.0
py==1.5.4
pyparsing==2.2.0
pytest==3.7.1
python-dateutil==2.7.3
pytz==2018.5
requests==2.18.4
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
tqdm==4.24.0
urllib3==1.22
Werkzeug==0.14.1
